requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-singularity==3.7.1